Simplify release.bat and update .gitignore
In release.bat, we remove the initial command lines "@echo off","chcp 65001", and "setlocal enabledelayedexpansion", and the terminal command "endlocal". This simplifies the build and deployment script, making it easily digestible and maintainable. Updates to the .gitignore file include removal of '*.iws', '*.iml', and '*.ipr' to clean the ignored patterns list, as these files are no longer relevant in our project structure.