Update Vert.x dependency version
https://github.com/vert-x3/wiki/wiki/4.4.6-Release-Notes

Bumped the Vert.x dependency version from 4.4.5 to 4.4.6 in pom.xml file. This was done to ensure our project is using the latest stable release which potentially includes bug fixes or enhancements that can improve the overall performance and reliability of our application. Regular dependency updates also help in keeping our codebase up-to-date with the latest standards.