Update project version and refactor RabbitMqHelper
The pom.xml file has been updated, moving from version 1.0.21 to 1.0.22, making ready for a new release. Additionally, a new file, TestCustomerHandlerTest.kt, was added in the test directory of the project to introduce testing for the TestCustomerHandler. Lastly, modifications were made in RabbitMqHelper.kt, where the assembleQueueName method was refactor. It's now assigning the exchangeType and queueName variables before appending them to queueName. This improves code readability and makes it easier to manage queue names.