Update project to version 1.0.25 and adjust code in RabbitMqHelper
Pom.xml files have been updated from version 1.0.24 to 1.0.25 for all modules within the project. This is part of a standard version increment to prepare for a new release. In RabbitMqHelper.kt, the surname of a exchange is now directly transformed to lowercase, improving readability and consistency within the code base.